Move backend auth types into a separate Type file
